backend/check_data.py
import mysql.connector
import os
from dotenv import load_dotenv

load_dotenv()

try:
    conn = mysql.connector.connect(
        host="localhost",
        user="root",
        password="",
        database="pelabuhan_db"
    )
    cursor = conn.cursor(dictionary=True)
    
    print("--- DATA SHIP ENTRIES ---")
    cursor.execute("SELECT id, nama_kapal, jenis_kegiatan, tanggal_kedatangan, MONTH(tanggal_kedatangan) as bln FROM ship_entries")
    rows = cursor.fetchall()
    for r in rows:
        print(r)
        
    conn.close()
except Exception as e:
    print(f"Error: {e}")
