backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
python-dotenv
apscheduler
openpyxl
pydantic
requests
pytz
